requirements.txt
amqp==5.0.9
APScheduler==3.7.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.3
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
crontab==0.23.0
cryptography==3.4.7
deep-diff==0.0.4
deepdiff==5.6.0
Deprecated==1.2.13
distlib==0.3.2
Django==3.2.5
django-cors-headers==3.7.0
django-crontab==0.7.1
django-redis==5.2.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.0.12
idna==3.3
ipynb-py-convert==0.4.6
kombu==5.2.3
lml==0.1.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.21.2
openpyxl==3.0.7
ordered-set==4.0.2
packaging==21.3
pandas==1.3.4
Pillow==8.3.2
pipreqs==0.4.11
platformdirs==2.3.0
prompt-toolkit==3.0.24
protobuf==3.19.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
pyenchant==3.2.2
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-xlsx==0.6.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
rejson==0.5.6
requests==2.26.0
schedule==1.1.0
six==1.16.0
somewhere==0.0.1
soupsieve==2.3.1
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.4
twilio==7.5.1
tzlocal==2.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.8.0
wcwidth==0.2.5
wrapt==1.13.3
XlsxWriter==3.0.3
yarg==0.1.9
